Refactor out ProcessAccountRequestPanelModule
