+++ src/web/app/components/process-account-request-panel/process-account-request-panel.component.scss
