+++ src/web/app/pages-admin/admin-requests-page/admin-requests-page.module.ts
@@ -3,15 +3,12 @@ import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { RouterModule, Routes } from '@angular/router';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
-import { AjaxLoadingModule } from '../../components/ajax-loading/ajax-loading.module';
 import { LoadingRetryModule } from '../../components/loading-retry/loading-retry.module';
 import { LoadingSpinnerModule } from '../../components/loading-spinner/loading-spinner.module';
-import { PanelChevronModule } from '../../components/panel-chevron/panel-chevron.module';
-import { TeammatesCommonModule } from '../../components/teammates-common/teammates-common.module';
-import { AdminRequestsPageComponent } from './admin-requests-page.component';
 import {
-  ProcessAccountRequestPanelComponent,
-} from './process-account-request-panel/process-account-request-panel.component';
+  ProcessAccountRequestPanelModule,
+} from '../../components/process-account-request-panel/process-account-request-panel.module';
+import { AdminRequestsPageComponent } from './admin-requests-page.component';
 
 const routes: Routes = [
   {
@@ -26,7 +23,6 @@ const routes: Routes = [
 @NgModule({
   declarations: [
     AdminRequestsPageComponent,
-    ProcessAccountRequestPanelComponent,
   ],
   exports: [
     AdminRequestsPageComponent,
@@ -34,13 +30,11 @@ const routes: Routes = [
   imports: [
     CommonModule,
     RouterModule.forChild(routes),
-    PanelChevronModule,
     LoadingRetryModule,
-    TeammatesCommonModule,
     LoadingSpinnerModule,
     FormsModule,
-    AjaxLoadingModule,
     NgbModule,
+    ProcessAccountRequestPanelModule,
   ],
 })
 export class AdminRequestsPageModule { }
